+++ admin/views/entities/delete/delete.js
@@ -1,8 +1,0 @@
-Template.adminEntitiesDelete.helpers({
-	onSuccess: function () {
-		return function (result) { 
-			var name = Router.current().data().entity.name;
-			Router.go('adminEntitiesIndex', {entity: name}); 
-		};
-	}
-});