+++ admin/views/users/delete/delete.js
@@ -1,6 +1,0 @@
-Template.adminUsersDelete.events({
-	'click .btn-delete': function () {
-		Meteor.users.remove(this._id);
-		Router.go('adminUsersIndex'); 
-	}
-});